requirements.txt.txt
feedparser
jinja2
pyyaml
transformers
torch
pillow
scipy
